♻️ refactor: rename and organize files
