+++ src/components/modal/ModalGlobal.tsx
@@ -4,7 +4,7 @@ import { Box, Modal } from "@mui/material";
 import ModalDeleteConfirm from "./ModalDeleteConfirm";
 import { ModalCreateAlert, ModalDeleteAlert } from "./ModalAlert";
 import ModalRepoDetails from "./ModalRepoDetails";
-import buildOptionsState from "lib/recoil/userBuildOptions";
+import { buildOptionsState } from "lib/recoil/buildOptions";
 import useModal from "lib/hooks/useModal";
 
 import { modalState } from "types/modal";
